pyproject.toml
[tool.poetry]
name = "demo-django"
version = "0.1.0"
description = ""
authors = ["wwfyde <wwfyde@163.com>"]
readme = "README.md"
#packages = [{include = "demo_django"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
pip = "^23.2.1"
mysqlclient = "^2.2.0"
pytest = "^7.4.2"
celery = "^5.3.4"
ipython = "^8.16.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
